VLC Plugin work in progress, misc fixes
1. A VLC Widget plugin is added, with widget commands to play and stop streams.

Video playback is not currently working but is compiling and running without crashing.
Audio playback is also not working via the API, but may "just work" through the device's native audio APIs.  (If so this would be a temporary solution until the Shellspace API gets audio)

2. The libvlc library is compiled into the core application.

The prebuilt library that is checked in was built according to the VLC for Android build instructions, taking the various library files (.a) from the intermediate build, and not the libvlcjni.so which has dependencies on a particular JNI interface and is intended to be used from Java, not C++.

Modifications were made, currently undocumented, to the VLC for Android sources to use stlport_static (instead of GNU STL) to stay in line with the rest of Shellspace.
In the future, the VLC plugin will be built as a separate shared object file and will not have to make the same STL choice as the Shellspace core.  These modifications included several hacks to the build scripts, the protobuf sources, and disabling the iomx plugin.

3. Gaze mouse is provisionally reimplemented (untested but compiles), and is now implemented in shell.js instead of the VNC widget sources.

This is intended allow other 2D widgets to take advantage of gaze mouse via the shell.

4. The menu now loads context-specific .menu files from the assets folder, instead of hardcoding them into menu.js.

5. Crittercism disabled for development.  Enabling Crittercism needs to become part of release.sh.

6. Adds missing "Connection: close" header to HTTP requests.

7. Fixes a minor issue with double '/' characters in file paths.

8. Adds a thread (copied from LibVLC Android source) which pipes stdout and stderr to the Android logging mechanism.

9. Cleans up unused VLC profilers.